requirements.txt
googlenewsdecoder
streamlit
google-generativeai
feedparser
python-dateutil
pytz